+++ include/SchoolClass.h
@@ -11,8 +11,8 @@ struct SchoolClass
         struct Node* next;
     } *head;
 
-    int num_of_students;
-    int class_number;
+    size_t num_of_students;
+    size_t class_number;
 };
 
 #endif
